♻️ refactor: useChatStore 경로 변경 및 파일들 경로 수정
